+++ master/buildbot/test/unit/process/test_buildstep.py
@@ -38,15 +38,15 @@ from buildbot.process.results import RETRY
 from buildbot.process.results import SKIPPED
 from buildbot.process.results import SUCCESS
 from buildbot.process.results import WARNINGS
-from buildbot.test.expect import ExpectGlob
-from buildbot.test.expect import ExpectMkdir
-from buildbot.test.expect import ExpectRmdir
-from buildbot.test.expect import ExpectShell
-from buildbot.test.expect import ExpectStat
 from buildbot.test.fake import fakebuild
 from buildbot.test.fake import fakemaster
 from buildbot.test.fake import worker
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectGlob
+from buildbot.test.steps import ExpectMkdir
+from buildbot.test.steps import ExpectRmdir
+from buildbot.test.steps import ExpectShell
+from buildbot.test.steps import ExpectStat
 from buildbot.test.steps import TestBuildStepMixin
 from buildbot.test.util import config
 from buildbot.test.util import interfaces
