+++ master/buildbot/test/unit/steps/test_package_deb_lintian.py
@@ -18,8 +18,8 @@ from twisted.trial import unittest
 from buildbot import config
 from buildbot.process.results import SUCCESS
 from buildbot.steps.package.deb import lintian
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
