+++ master/buildbot/test/unit/steps/test_package_rpm_mock.py
@@ -19,9 +19,9 @@ from buildbot import config
 from buildbot.process.properties import Interpolate
 from buildbot.process.results import SUCCESS
 from buildbot.steps.package.rpm import mock
-from buildbot.test.expect import ExpectRmdir
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectRmdir
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
