+++ master/buildbot/test/unit/steps/test_shellsequence.py
@@ -22,8 +22,8 @@ from buildbot.process.results import FAILURE
 from buildbot.process.results import SUCCESS
 from buildbot.process.results import WARNINGS
 from buildbot.steps import shellsequence
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
 from buildbot.test.util import config as configmixin
 from buildbot.test.util.warnings import assertProducesWarnings
