+++ master/buildbot/test/unit/steps/test_source_git.py
@@ -27,15 +27,15 @@ from buildbot.process.results import FAILURE
 from buildbot.process.results import RETRY
 from buildbot.process.results import SUCCESS
 from buildbot.steps.source import git
-from buildbot.test.expect import ExpectCpdir
-from buildbot.test.expect import ExpectDownloadFile
-from buildbot.test.expect import ExpectListdir
-from buildbot.test.expect import ExpectMkdir
-from buildbot.test.expect import ExpectRemoteRef
-from buildbot.test.expect import ExpectRmdir
-from buildbot.test.expect import ExpectShell
-from buildbot.test.expect import ExpectStat
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectCpdir
+from buildbot.test.steps import ExpectDownloadFile
+from buildbot.test.steps import ExpectListdir
+from buildbot.test.steps import ExpectMkdir
+from buildbot.test.steps import ExpectRemoteRef
+from buildbot.test.steps import ExpectRmdir
+from buildbot.test.steps import ExpectShell
+from buildbot.test.steps import ExpectStat
 from buildbot.test.steps import TestBuildStepMixin
 from buildbot.test.util import config
 from buildbot.test.util import sourcesteps
