+++ master/buildbot/test/unit/steps/test_source_p4.py
@@ -24,9 +24,9 @@ from twisted.trial import unittest
 from buildbot.process.results import RETRY
 from buildbot.process.results import SUCCESS
 from buildbot.steps.source.p4 import P4
-from buildbot.test.expect import ExpectRmdir
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectRmdir
+from buildbot.test.steps import ExpectShell
 from buildbot.test.util import sourcesteps
 from buildbot.test.util.config import ConfigErrorsMixin
 from buildbot.test.util.properties import ConstantRenderable
