+++ master/buildbot/test/unit/steps/test_subunit.py
@@ -22,8 +22,8 @@ from twisted.trial import unittest
 from buildbot.process.results import FAILURE
 from buildbot.process.results import SUCCESS
 from buildbot.steps import subunit
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
 
 try:
