+++ master/buildbot/test/unit/steps/test_vstudio.py
@@ -25,8 +25,8 @@ from buildbot.process.results import FAILURE
 from buildbot.process.results import SUCCESS
 from buildbot.process.results import WARNINGS
 from buildbot.steps import vstudio
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
 
 real_log = r"""
