+++ master/buildbot/test/unit/test_steps_git_diffinfo.py
@@ -17,8 +17,8 @@ from twisted.trial import unittest
 
 from buildbot.process import results
 from buildbot.steps import gitdiffinfo
-from buildbot.test.expect import ExpectShell
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.steps import ExpectShell
 from buildbot.test.steps import TestBuildStepMixin
 
 try:
